move struct definition to header
